backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
motor>=3.1.1
pymongo>=4.3.3
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
email-validator>=2.0.0 